+++ Src/DependencyCollector/Shared/Implementation/ProfilerHttpProcessing.cs
@@ -2,13 +2,7 @@
 {
     using System;
     using System.Collections.Generic;
-    using System.Diagnostics;
-    using System.Globalization;
-    using System.Linq;
     using System.Net;
-    using System.Web;
-    using Extensibility.Implementation.Tracing;
-    using Microsoft.ApplicationInsights.Common;
     using Microsoft.ApplicationInsights.DataContracts;
     using Microsoft.ApplicationInsights.DependencyCollector.Implementation.Operation;
     using Microsoft.ApplicationInsights.Extensibility;
@@ -57,7 +51,7 @@
         /// <returns>The resulting return value.</returns>
         public object OnEndForGetResponse(object context, object returnValue, object thisObj)
         {
-            this.OnEnd(null, thisObj, returnValue);
+            this.OnEndResponse(thisObj, returnValue);
             return returnValue;
         }
 
@@ -69,7 +63,7 @@
         /// <param name="thisObj">This object.</param>        
         public void OnExceptionForGetResponse(object context, object exception, object thisObj)
         {
-            this.OnEnd(exception, thisObj, null);
+            this.OnEndException(exception, thisObj);
         }
         
         /// <summary>
@@ -93,7 +87,7 @@
         /// <param name="transportContext">The transport context parameter.</param>
         public void OnExceptionForGetRequestStream(object context, object exception, object thisObj, object transportContext)
         {
-            this.OnEnd(exception, thisObj, null);
+            this.OnEndException(exception, thisObj);
         }
 
         /// <summary>
@@ -118,7 +112,7 @@
         /// <returns>The return value passed.</returns>
         public object OnEndForEndGetResponse(object context, object returnValue, object thisObj, object asyncResult)
         {
-            this.OnEnd(null, thisObj, returnValue);
+            this.OnEndResponse(thisObj, returnValue);
             return returnValue;
         }
 
@@ -131,7 +125,7 @@
         /// <param name="asyncResult">The asyncResult parameter.</param>
         public void OnExceptionForEndGetResponse(object context, object exception, object thisObj, object asyncResult)
         {
-            this.OnEnd(exception, thisObj, null);
+            this.OnEndException(exception, thisObj);
         }
 
         /// <summary>
@@ -157,7 +151,7 @@
         /// <param name="transportContext">The transportContext parameter.</param>
         public void OnExceptionForEndGetRequestStream(object context, object exception, object thisObj, object asyncResult, object transportContext)
         {
-            this.OnEnd(exception, thisObj, null);
+            this.OnEndException(exception, thisObj);
         }
 
         #endregion // Http callbacks