+++ Test/PerformanceCollector/FunctionalTests/Test45.cs
@@ -18,7 +18,7 @@
         private const int TimeoutInMs = 15000;
 
         private const string TestWebApplicationSourcePath = @"TestApps\TestApp45\App";
-        private const string TestWebApplicationDestPath = @"TestsPerformanceCollector45";
+        private const string TestWebApplicationDestPath = @"TestApps\TestApp45\App";
 
         [TestInitialize]
         public void TestInitialize()
@@ -30,7 +30,8 @@
             var applicationDirectory = Path.Combine(
                 Directory.GetCurrentDirectory(),
                 string.Format("{0}_{1}", "app", this.TestContext.TestName));
-            
+            applicationDirectory = Path.GetFullPath(applicationDirectory);
+
             CopyFolder(originalDirectory, applicationDirectory);
 
             Trace.WriteLine("Application directory:" + applicationDirectory);
@@ -43,6 +44,12 @@
             int telemetryPort = rand.Next(minPort, maxPort + 1);
             int quickPulsePort = rand.Next(minPort, maxPort + 1);
 
+            Trace.WriteLine("IIS Port:" + iisPort);
+            Trace.WriteLine("Telemetry Port:" + telemetryPort);
+            Trace.WriteLine("Quickpulse Port:" + quickPulsePort);
+
+           
+            Trace.WriteLine(DateTime.UtcNow.ToLongTimeString() + ":Starting Web host");
             this.StartWebAppHost(
                 new SingleWebHostTestConfiguration(
                     new IisExpressConfiguration
@@ -57,13 +64,15 @@
                     AttachDebugger = Debugger.IsAttached,
                     IKey = "fafa4b10-03d3-4bb0-98f4-364f0bdf5df8",
                 });
-
             OverwriteFile(applicationDirectory, "ApplicationInsights.config", new Dictionary<string, string>() { ["{TelemetryEndpointPort}"] = telemetryPort.ToString() });
             OverwriteFile(applicationDirectory, "Web.config", new Dictionary<string, string>() { ["{QuickPulseEndpointPort}"] = quickPulsePort.ToString() });
 
+            Trace.WriteLine(DateTime.UtcNow.ToLongTimeString() + ":Starting Web host completed successfully");
             try
             {
+                Trace.WriteLine(DateTime.UtcNow.ToLongTimeString() + ": Launching application and verifying..");
                 base.LaunchAndVerifyApplication();
+                Trace.WriteLine(DateTime.UtcNow.ToLongTimeString() + ": Launching application and verifying success.");
             }
             catch (Exception ex)
             {
@@ -86,6 +95,8 @@
             }
 
             File.WriteAllText(filePath, fileContent);
+
+            Trace.WriteLine(DateTime.UtcNow.ToLongTimeString() + ": FileOverWritten: " + filePath);
         }
 
         [TestCleanup]
@@ -94,67 +105,51 @@
             this.StopWebAppHost();
         }
 
-        [TestMethod]
-        [Owner("alkaplan")]
-        [DeploymentItem(TestWebApplicationSourcePath, TestWebApplicationDestPath)]
+        [TestMethod]        
         public void DefaultCounterCollection()
         {
             CommonTests.DefaultCounterCollection(this.Listener);
         }
 
-        [TestMethod]
-        [Owner("alkaplan")]
-        [DeploymentItem(TestWebApplicationSourcePath, TestWebApplicationDestPath)]
+        [TestMethod]        
         public void CustomCounterCollection()
         {
             CommonTests.CustomCounterCollection(this.Listener);
         }
 
-        [TestMethod]
-        [Owner("alkaplan")]
-        [Description("Tests that non existent counters are not collected and wont affect other counters")]
-        [DeploymentItem(TestWebApplicationSourcePath, TestWebApplicationDestPath)]
+        [TestMethod]        
+        [Description("Tests that non existent counters are not collected and wont affect other counters")]        
         public void NonExistentCounter()
         {
             CommonTests.NonExistentCounter(this.Listener);
         }
 
-        [TestMethod]
-        [Owner("alkaplan")]
-        [Description("Tests that non existent counters which use placeholders are not collected and wont affect other counters")]
-        [DeploymentItem(TestWebApplicationSourcePath, TestWebApplicationDestPath)]
+        [TestMethod]        
+        [Description("Tests that non existent counters which use placeholders are not collected and wont affect other counters")]        
         public void NonExistentCounterWhichUsesPlaceHolder()
         {
             CommonTests.NonExistentCounterWhichUsesPlaceHolder(this.Listener);
         }
 
-        [TestMethod]
-        [Owner("alkaplan")]
-        [DeploymentItem(TestWebApplicationSourcePath, TestWebApplicationDestPath)]
+        [TestMethod]           
         public void NonParsableCounter()
         {
             CommonTests.NonParsableCounter(this.Listener);
         }
 
-        [TestMethod]
-        [Owner("alkaplan")]
-        [DeploymentItem(TestWebApplicationSourcePath, TestWebApplicationDestPath)]
+        [TestMethod]           
         public void QuickPulseAggregates()
         {
             CommonTests.QuickPulseAggregates(this.QuickPulseListener, this.HttpClient);
         }
 
-        [TestMethod]
-        [Owner("alkaplan")]
-        [DeploymentItem(TestWebApplicationSourcePath, TestWebApplicationDestPath)]
+        [TestMethod]             
         public void QuickPulseMetricsAndDocuments()
         {
             CommonTests.QuickPulseMetricsAndDocuments(this.QuickPulseListener, this);
         }
 
-        [TestMethod]
-        [Owner("alkaplan")]
-        [DeploymentItem(TestWebApplicationSourcePath, TestWebApplicationDestPath)]
+        [TestMethod]        
         public void QuickPulseTopCpuProcesses()
         {
             CommonTests.QuickPulseTopCpuProcesses(this.QuickPulseListener, this);